+++ changelog.md
@@ -1,2 +1,3 @@
-### Kube v0.1.2 release
-Fix box behavior
+### Kube v0.1.6 release
+- Fixed box child behavior
+- Added playground webpage